Refactor QC reporting & move report-specific code into qc/reporting.py.
